Price-Quote Application Assignment Submission
Your job is to create a console-based application for getting a shipping quote for a package 
Requirements

The first line of the program must be: “Welcome to Package Express. Please follow the instructions below.”
The user must then be prompted for the package weight.
If the weight is greater than 50, display the error message, “Package too heavy to be shipped via Package Express. Have a good day.” At this point the program will end.
The user must then be prompted for:
the package width,
then the package height,
then the package length.
If the total of all dimensions is greater than 50, display the error message, “Package too big to be shipped via Package Express.” At this point the program will end.
Next, multiply the three dimensions (height, width, & length) together, and multiply the product by the weight. Finally, divide the outcome by 100.
The result of this calculation is the price quote.
Display the quote to the user as a dollar amount.